cleanup

+++ source/pipelines/custom/TexturedPlane.cpp
@@ -85,7 +85,7 @@ void TexturedPlane::createDescriptorSets()
   const VkDescriptorSetAllocateInfo allocateInfo {
     .sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO,
     .descriptorPool = descriptorPool,
-    .descriptorSetCount = static_cast<uint32_t>(logicalDevice->getMaxFramesInFlight()),
+    .descriptorSetCount = logicalDevice->getMaxFramesInFlight(),
     .pSetLayouts = layouts.data()
   };
